moved markdown parser file to src folder
