+++ integrations/line/src/index.ts
@@ -2,7 +2,7 @@ import { RuntimeError } from '@botpress/client'
 import { sentry as sentryHelpers } from '@botpress/sdk-addons'
 import { messagingApi as lineMessagingApi } from '@line/bot-sdk'
 import crypto from 'crypto'
-import { parseMarkdown } from 'markdown-parser'
+import { parseMarkdown } from './markdown-parser'
 import getOrCreateConversation from './proactive-conversation'
 import getOrCreateUser from './proactive-user'
 import * as bp from '.botpress'
